+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * [#984](https://github.com/mozilla/glean.js/pull/984): BUGFIX: Return correct upload result in case an error happens while building a ping request.
 * [#988](https://github.com/mozilla/glean.js/pull/988): BUGFIX: Enforce rate limitation at upload time, not at ping submission time.
+* [#1006](https://github.com/mozilla/glean.js/pull/1006): Implement the rate metric.
+
   * Note: This change required a big refactoring of the internal uploading logic.
 
 # v0.27.0 (2021-11-22)